remove qubit/parametric_ops_controlled.py
