+++ pennylane/ops/qubit/__init__.py
@@ -24,7 +24,6 @@ The operations are divided into the following files:
   located in ``non_parameteric_ops.py`` instead.
 * ``parametric_ops_single_qubit.py``: Core single qubit parametric operations.
 * ``parametric_ops_multi_qubit.py``: Core multi-qubit parametric operations.
-* ``parametric_ops_controlled.py``: Controlled parametric operations.
 * ``qchem_ops.py``: Operations for quantum chemistry applications.
 * ``state_preparation.py``: Operations that initialize the state.
 * ``special_unitary.py``: The ``SpecialUnitary`` operation.
@@ -36,7 +35,6 @@ from .non_parametric_ops import *
 from .observables import *
 from .parametric_ops_single_qubit import *
 from .parametric_ops_multi_qubit import *
-from .parametric_ops_controlled import *
 from .qchem_ops import *
 from .state_preparation import *
 from .special_unitary import SpecialUnitary
